refactor(i18n): edit chinese and indonesian ISO 639-1 codes
