+++ src/locales/README.md
@@ -1,0 +1,31 @@
+# Codes to use to identify languages
+
+The [IETF language tags](http://en.wikipedia.org/wiki/IETF_language_tag) are made up of [ISO 639](http://en.wikipedia.org/wiki/ISO_639) standards for representing language names and [ISO 3166-1](http://en.wikipedia.org/wiki/ISO_3166-1) standards for representing country codes. A sample list:
+
+| English name for language | Tag   |
+| :------------------------ | :---- |
+| English                   | en    |
+| English (United States)   | en-US |
+| English (Great Britain)   | en-GB |
+| French                    | fr    |
+| German                    | de    |
+| Polish                    | pl    |
+| Dutch                     | nl    |
+| Finnish                   | fi    |
+| Swedish                   | sv    |
+| Italian                   | it    |
+| Spanish (Spain)           | es    |
+| Portuguese (Portugal)     | pt    |
+| Russian                   | ru    |
+| Portuguese (Brazil)       | pt-BR |
+| Spanish (Mexico)          | es-MX |
+| Chinese (PRC)             | zh-CN |
+| Chinese (Taiwan)          | zh-TW |
+| Japanese                  | ja    |
+| Korean                    | ko    |
+
+Use these two reference lists to create new languages.
+
+[ISO 639-1](https://datahub.io/core/language-codes/r/0.html)
+
+[IETF Language Types](https://datahub.io/core/language-codes/r/3.html)
