+++ src/locales/id/translations/actions/index.ts
