+++ src/locales/id/translations/dashboard/index.ts
