+++ src/locales/id/translations/index.ts
