+++ src/locales/id/translations/labs/index.ts
