+++ src/locales/id/translations/patient/index.ts
