+++ src/locales/id/translations/patients/index.ts
