+++ src/locales/zhCN/translations/actions/index.ts
