+++ src/locales/zhCN/translations/dashboard/index.ts
