+++ src/locales/zhCN/translations/index.ts
