+++ src/locales/zhCN/translations/labs/index.ts
