+++ src/locales/zhCN/translations/patient/index.ts
