+++ src/locales/zhCN/translations/patients/index.ts
@@ -1,6 +1,6 @@
 export default {
   patients: {
-    label: '耐心',
+    label: '标签',
     viewPatients: '查看患者',
     viewPatient: '查看患者',
     newPatient: '新病人',
